requirements.txt
# 核心依赖
oemer>=0.1.5
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# MIDI处理
mido>=1.3.0
pretty-midi>=0.2.10
music21>=9.1.0

# 图像处理
scikit-image>=0.21.0
matplotlib>=3.7.0

# Web框架
Flask>=2.3.0
Flask-CORS>=4.0.0
gunicorn>=21.0.0

# 文件处理
python-multipart>=0.0.6
Werkzeug>=2.3.0

# 进度和日志
tqdm>=4.65.0
loguru>=0.7.0

# 测试
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# 开发工具
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# 可选依赖 (GPU加速)
# torch>=2.0.0
# torchvision>=0.15.0
# onnxruntime-gpu>=1.15.0

# 乐谱渲染 (需要额外安装LilyPond)
# abjad>=3.18
